+++ client/App.tsx
@@ -1,7 +1,9 @@
 import { Route, BrowserRouter, Routes } from "react-router-dom"
-import { HomePage } from "./pages/home/home-page"
-import { EditorPage } from "./pages/editor/editor-page"
 import { NetworkManager } from "./Network/manager"
+
+import { LoginPage } from "./pages/login-page/login-page"
+import { EditorPage } from "./pages/editor/editor-page"
+import { HomePage } from "./pages/home/home-page"
 
 function App() {
     // Reconnect to the server each time a state changes
@@ -10,7 +12,8 @@ function App() {
     return (    
         <BrowserRouter>
             <Routes>
-                <Route path="/" element={<HomePage />} />
+                <Route path="/" element={<LoginPage />} />
+                <Route path="/home" element={<HomePage />} />
                 <Route path="/editor" element={<EditorPage />} />
             </Routes>
         </BrowserRouter>
