+++ client/main.css
@@ -2,6 +2,9 @@
 body, html {
     width: 100%;
     height: 100%;
+    margin: 0;
+    padding: 0;
+    background-color: #1c1c1b;
 }
 
 #root {
